test: Update tests with fix for NotificationContentRepository
