+++ src/test/java/com/synopsys/integration/alert/workflow/message/NotificationReceiverTestIT.java
@@ -7,9 +7,7 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 import java.util.ArrayList;
 import java.util.List;
 
-import org.junit.Ignore;
 import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.PageRequest;
@@ -41,10 +39,7 @@ public class NotificationReceiverTestIT {
         notificationContentRepository.deleteAllInBatch();
     }
 
-    //FIXME: This test will now fail due to the @Transactional annotation, once it is removed it should be renabled
     @Test
-    @Ignore
-    @Disabled
     public void testHandleEventNotProcessedNotifications() {
         List<AlertNotificationModel> notificationContent = new ArrayList<>();
         notificationContent.add(createAlertNotificationModel(1L, false));
@@ -73,10 +68,7 @@ public class NotificationReceiverTestIT {
         testAlertNotificationModels(savedModels);
     }
 
-    //FIXME: This test will now fail due to the @Transactional annotation, once it is removed it should be renabled
     @Test
-    @Ignore
-    @Disabled
     public void testHandleEventMixedProcessedNotifications() {
         List<AlertNotificationModel> notificationContent = new ArrayList<>();
         notificationContent.add(createAlertNotificationModel(1L, true));
